Add files via upload
The JIT Compiled Expression Evaluator is a program that takes a mathematical expression as input, dynamically generates C code for it, compiles the code at runtime, and executes it to evaluate the expression. This process, known as Just-In-Time (JIT) compilation, allows for optimized execution by compiling the code on the fly. The project also includes an extra feature to implement a sigmoid function